[FEAT] #75 uuid 온보딩으로 넘김
