+++ app/src/main/java/com/timi/seulseul/presentation/MainViewModel.kt
@@ -30,39 +30,14 @@ import javax.inject.Inject
 // Hilt가 ViewModel 객체를 생성하고 관리할 수 있게 해준다.
 @HiltViewModel
 class MainViewModel @Inject constructor(
-    private val authRepo: AuthRepo,
     private val alarmRepo : AlarmRepo,
     private val subwayRouteRepo: SubwayRouteRepo
 ) : ViewModel() {
-
-    // Result<AuthData>를 _auth에 저장한다. 내부에서만 수정 가능
-    private var _auth = MutableLiveData<Result<AuthData>>()
-    // 외부에서 접근 가능한 auth는 읽기만 된다.
-    var auth : LiveData<Result<AuthData>> = _auth
-
     private var _subwayData : MutableLiveData<MutableList<SubwayRouteItem>> = MutableLiveData()
     var subwayData : LiveData<MutableList<SubwayRouteItem>> = _subwayData
 
     private var _showRoute : MutableLiveData<Boolean> = MutableLiveData(false)
     var showRoute : LiveData<Boolean> = _showRoute
-
-    // postAuth함수는
-    // uuid값을 넣은 Auth 객체를 만들고 authRepo.postAuth 함수에 전달한다.
-    fun postAuth() {
-        // uuid를 sp에서 가져온다.
-        var uuid = prefs.getString("KEY_UUID", null)
-
-        // 없다면 생성하고 sp에 저장한다.
-        if (uuid == null) {
-            uuid = UUID.randomUUID().toString()
-            prefs.edit().putString("KEY_UUID", uuid).apply()
-        }
-
-        viewModelScope.launch{
-            // authRepo에서 postAuth를 호출해 반환된 결과 값을 _auth LiveData에 저장한다.
-            _auth.value = authRepo.postAuth(Auth(uuid))
-        }
-    }
 
     fun setAlarm() {
         if (!prefs.getBoolean("isNotFirst", false)) {
